refactor: split up timekeep logic into separate files
