+++ src/timekeep/__fixtures__/manipulating/collapse/shouldNotCollapseSingleEntry.ts
@@ -1,0 +1,12 @@
+import moment from "moment";
+import { TimeEntry } from "@/timekeep/schema";
+
+export const currentTime = moment();
+
+export const input: TimeEntry = {
+	id: "9054dee3-8c15-493b-ad31-f070e08c2699",
+	name: "Test",
+	startTime: currentTime,
+	endTime: currentTime,
+	subEntries: null,
+};
